pyproject.toml
[tool.poetry]
name = "splot"
version = "0.1.0"
description = "Command line tool for managing your Spotify account."
authors = ["Niels Godfredsen <ngodfredsen@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
flake8 = "^3.8.4"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
